ArbitraryStyleTransfer/requirements.txt
numpy==1.17.2
Pillow==8.2.0
pkg-resources==0.0.0
protobuf==3.9.1
six==1.12.0
tensorboardX==1.8
torch==1.2.0
torchvision==0.4.0
tqdm==4.35.0
opencv-python==4.4.0.46
imageio==2.9.0
scikit-image
